Add throttling support for the AWS Logs and AWS Flow Logs inputs
This change applies only to the Kinesis transport which is used for these inputs. If the throttling state becomes active while processing messages, then the Kinesis consumer will wait up to 60 seconds (during which the input will continue processing if the throttling state clears). If the throttling state does not clear, then the consumer will be temporarily shut down until the throttling state is cleared.