Always load AWS plugin but guard inputs with leader check
To support a node becoming a leader at runtime, we have to load the
plugin on all nodes. Previously it was only loaded at the master node.

To prevent the AWS inputs to be started on a non-leader node, mark the
inputs as "only one per cluster", so that the input system can take care
of this automatically.